requirements.txt
streamlit~=1.25.0
pandas~=2.0.3
Pillow~=9.5.0
openpyxl~=3.1.2
python-docx~=0.8.11
SQLAlchemy~=2.0.19
pymysql~=1.1.0
PyYAML~=6.0.1